Seperation of subscription from processing.
4 threads to process.

Notification on first load for shiny app.
